@@ -2,7 +2,7 @@ package net.automotons.items.heads;
 
 import net.automotons.blocks.AutomotonBlockEntity;
 import net.automotons.items.HeadItem;
-import net.minecraft.nbt.CompoundTag;
+import net.minecraft.nbt.NbtCompound;
 import net.minecraft.util.math.Direction;
 
 public class RedstoneHeadItem extends HeadItem<Boolean>{
@@ -15,12 +15,12 @@ public class RedstoneHeadItem extends HeadItem<Boolean>{
 		return automoton.engaged ? (muted != null && muted) ? 1 : 15 : 0;
 	}
 	
-	public Boolean readExtraData(CompoundTag tag){
+	public Boolean readExtraData(NbtCompound tag){
 		return tag.getBoolean("muted");
 	}
 	
-	public CompoundTag writeExtraData(Boolean muted){
-		CompoundTag tag = new CompoundTag();
+	public NbtCompound writeExtraData(Boolean muted){
+		NbtCompound tag = new NbtCompound();
 		tag.putBoolean("muted", muted != null && muted);
 		return tag;
 	}
