@@ -3,6 +3,7 @@ package net.automotons.screens;
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.automotons.blocks.AutomotonBlockEntity;
 import net.minecraft.client.gui.screen.ingame.HandledScreen;
+import net.minecraft.client.render.GameRenderer;
 import net.minecraft.client.resource.language.I18n;
 import net.minecraft.client.util.math.MatrixStack;
 import net.minecraft.entity.player.PlayerInventory;
@@ -28,10 +29,10 @@ public class AutomotonScreen extends HandledScreen<AutomotonScreenHandler>{
 		playerInventoryTitleY = -200000;
 	}
 	
-	@SuppressWarnings("deprecation")
 	protected void drawBackground(MatrixStack matrices, float delta, int mouseX, int mouseY){
-		RenderSystem.color4f(1, 1, 1, 1);
-		client.getTextureManager().bindTexture(TEXTURE);
+		RenderSystem.setShader(GameRenderer::getPositionTexShader);
+		RenderSystem.setShaderColor(1.0F, 1.0F, 1.0F, 1.0F);
+		RenderSystem.setShaderTexture(0, TEXTURE);
 		int x = (width - backgroundWidth) / 2;
 		int y = (height - backgroundHeight) / 2;
 		drawTexture(matrices, x, y, 0, 0, backgroundWidth, backgroundHeight);
